Report init process errors during early startup
If the init process fails during StartInitialization() it returns the cause to the parent via the pipe.

This initial failure causes knock on failures in parent which are reported and the error from the init process is lost.

For the early portion of start() we now attempt to read the error from the init process, if one is found we report it in preference, as it provides more relevant information about the actual problem.

Signed-off-by: Steven Hartland <steven.hartland@multiplay.co.uk>